about project
A python project using the tkinter and pickle libraries. A project with authorization saved to a file. There are Toyota legacies such as the Camry and Land Cruiser. For each object there is a photo imported from the file to which the path is specified. The functions also include tuning the car, adding cars, as well as removing them. But before deleting the machine, the object is sent to the trash from where it is already possible to permanently delete the object.